Changed hud such that it now has a button handler to handle events
Top peek requires the card user to be able to move again after peeking. This requires the buttons to be re-enabled such that the playaer can move again. Because of this, the event listener mechanic is reworked to have a Future<String> method for waiting on the player's decision, and handle cases based on it